pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "music-finder"
version = "1.0.0"
description = "Niche artist discovery pipeline — finds emerging artists matching your taste"
requires-python = ">=3.9"
dependencies = [
    "spotipy>=2.23.0",
    "musicbrainzngs>=0.7.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "apscheduler>=3.10.4",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.hatch.build.targets.wheel]
packages = ["src/music_finder"]
